Refactor (contentType) : 시그니처 변경, 기본값 `HTML` 추가
